Finilized removal of vagrant public key from image
The vagrant public key was preserved and restored on every boot
if no the context variable SSH_PUBLIC_KEY was missing.
The initial install.yml file and the vagrant public key
are stored under /var/lib/coreos-install/user_data. This files gets
applied on every reboot.
To bring in the public key for provisioning the following trick
is used:
By adding the custom qemu arguments a virtual disk is mounted
and processed by CoreOS after the first boot.
https://coreos.com/os/docs/latest/config-drive.html

Use the environment variable REMOVE_VAGRANT_KEY to adjust the
removal.